+++ src/hooks/DataTable.ts
@@ -1,9 +1,31 @@
-import { nextTick, reactive, ref, shallowReactive } from 'vue'
+import { nextTick, reactive, Ref, ref } from 'vue'
 import useEmit from './Emit'
 
-export default function (): Record<string, any> {
-  const dataList = shallowReactive([]) as Array<any>
-  const selectRows = ref<Array<any>>([])
+export interface IDataTable<T> {
+  dataList: Ref<T[] | undefined>
+  selectRows: Ref<T[] | undefined>
+  tableConfig: {
+    stripe: boolean
+    border: boolean
+    size: string
+    headerCellStyle: any
+    height: number
+  }
+  tableLoading: Ref<boolean>
+  handleSuccess: ({
+    data,
+    totalSize,
+  }: {
+    data: T[]
+    totalSize?: number
+  }) => Promise<{ data: T[] }>
+  handleSelectionChange: (tempSelectRows: T[]) => void
+  [key: string]: any
+}
+
+export default function <T>(): IDataTable<T> {
+  const dataList = ref<T[]>()
+  const selectRows = ref<T[]>()
   const tableConfig = reactive({
     stripe: true,
     border: false,
@@ -14,13 +36,18 @@ export default function (): Record<string, any> {
     height: 200,
   })
   const tableLoading = ref(true)
-  const handleSuccess = ({ data = [], totalSize = 10 }) => {
-    dataList.length = 0
-    dataList.push(...data)
+  const handleSuccess = ({
+    data = [],
+    totalSize = 10,
+  }: {
+    data: T[]
+    totalSize?: number
+  }) => {
+    dataList.value = data
     tableLoading.value = false
     return Promise.resolve({ data, totalSize })
   }
-  const handleSelectionChange = (tempSelectRows: Array<any>) => {
+  const handleSelectionChange = (tempSelectRows: T[]) => {
     selectRows.value = tempSelectRows
   }
   const pageChanged = () => {
@@ -77,8 +104,7 @@ export default function (): Record<string, any> {
     tableLoading,
     handleSuccess,
     handleSelectionChange,
-    doRefresh,
-    useHeight,
     offTableCollapseTransition,
+    useHeight,
   }
 }